Move test of array_inputs strategies into test_strategies module.
As well as needed utility functions into util module.
