+++ python/cuml/tests/test_strategies.py
@@ -13,14 +13,41 @@
 # limitations under the License.
 #
 from cuml.testing.strategies import (
-    standard_regression_datasets,
+    create_cuml_array_input,
+    cuml_array_dtypes,
+    cuml_array_input_types,
+    cuml_array_orders,
+    cuml_array_shapes,
     regression_datasets,
     split_datasets,
     standard_datasets,
+    standard_regression_datasets,
 )
+from cuml.testing.utils import normalized_shape, series_squeezed_shape
 from hypothesis import given, settings, HealthCheck
 from hypothesis import strategies as st
 from hypothesis.extra.numpy import floating_dtypes
+
+
+@given(
+    input_type=cuml_array_input_types(),
+    dtype=cuml_array_dtypes(),
+    shape=cuml_array_shapes(),
+    order=cuml_array_orders())
+@settings(deadline=None)
+def test_array_inputs(input_type, dtype, shape, order):
+    input_array = create_cuml_array_input(input_type, dtype, shape, order)
+    assert input_array.dtype == dtype
+    if input_type == "series":
+        assert input_array.shape == series_squeezed_shape(shape)
+    else:
+        assert input_array.shape == normalized_shape(shape)
+
+    layout_flag = f"{order}_CONTIGUOUS"
+    if input_type == "series":
+        assert input_array.values.flags[layout_flag]
+    else:
+        assert input_array.flags[layout_flag]
 
 
 @given(standard_datasets())
